+++ e2etest/workspace_test.go
@@ -7,43 +7,19 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func init() {
-	/*
-	   program := "kubectl"
-	   appcmd := "apply"
-	   flags := "-f"
-	   arg := "workspace_cm_test.yaml"
+func (s *e2eTestSuite) Test_Workspaces() {
+	requestor := make_authenticated_requestor(s.client, mockUser)
 
-	   	if exec.Command(program, appcmd, flags, arg).Run() != nil {
-	   		panic("Error applying workspace manifests")
-	   	}
-	*/
-}
-
-func (s *e2eTestSuite) Test_workspaces() {
-	requestor := make_requestor(s.client)
-
-	resp, err := requestor("http://localhost:8842/api/v1/workspaces/", http.MethodGet, "")
-	require.NoError(s.T(), err, "no point continue")
-	require.Equal(s.T(), http.StatusOK, resp.StatusCode)
-
-	s.Len(resp.Header["Content-Type"], 1)
-	s.Equal("application/json", resp.Header["Content-Type"][0])
+	resp, err := requestor(server_addr+"/api/v1/workspaces/", http.MethodGet, "")
+	require.NoError(s.T(), err, BodyStringer{resp.Body})
+	require.Equal(s.T(), http.StatusOK, resp.StatusCode, BodyStringer{resp.Body})
 
 	type WorkspaceList struct {
 		Items []workspace.Workspace `json:"items"`
 	}
 	var list WorkspaceList
-	marshalResponse(ResponseBodyBytes(resp), &list)
+	err = marshalResponse(ResponseBodyBytes(resp), &list)
 
-	s.Len(list.Items, 2)
-
-	accesses := make([]bool, 2)
-
-	for i, ws := range list.Items {
-		accesses[i] = ws.HasAccess
-	}
-
-	s.Contains(accesses, false)
-	s.Contains(accesses, true)
+	s.NoError(err)
+	s.NotEmpty(list.Items)
 }
